pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cyber-agent"
version = "0.1.0"
description = "Protego — AI threat defense multi-agent app for small business (watsonx.ai + LangGraph)"
requires-python = ">=3.10"
dependencies = [
    "langgraph>=0.2.50",
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-ibm>=0.3.0",
    "langchain-community>=0.3.0",
    "pydantic>=2.6",
    "python-dotenv>=1.0",
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "python-multipart>=0.0.9",
    "pdfplumber>=0.11",
    "pytesseract>=0.3.10",
    "Pillow>=10.0",
    "httpx>=0.27",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
